Remove unnecessary catch for file not found
We're already handling this error for both Prelude.interfaces and normal elmi
file deserialization, so we no longer need to catch an IO error
here. InterfaceSerialization.loadInterface returns an IO (Either String
L.ByteString) value that contains an error message in case the file is not
found, so I don't think this catch is going to trigger any more.