Add basic Travis CI configuration
I added a basic Travis CI configuration to make the build status more
transparent. I added a small logo to the bottom of the README with the
status and a link to the build.

@evancz after merging this you'd have to go to https://travis-ci.org
and enable the build for this project, just takes a couple of clicks.
See http://about.travis-ci.org/docs/user/getting-started/.

I tested this on my own github account and the configuration worked
(ie, it showed the same build errors on Travis that I'm experiencing
locally).