Get rid of Selection, simplifying the field API
Real reason is that setSelectionRange also sets the focus. This
behavior makes sense, so the underlying issue is that setting the
selection for two fields is not meaningful because only one can have
focus.

How to set focus (for fields or inputs in general) is not really
answered yet, so I think it makes the most sense to leave it out for
now.