Remove Array for now, migrate to vector branch
I got around to properly doing my research on various implementations
for immutable vectors. It is not clear to me that the RRB-Trees are
going to be the best trade-off for the typical uses seen in Elm, so I
want to try the Clojure approach as well and race them.