Drop higher kinded polymorphism (hkp) for now
Do not want the community to get distracted trying to be like Haskell
when the goal is to be better than JavaScript.

Articles about making an app that looks great are positive for Elm!
Articles about how simple it is to architect and maintain an Elm app
are great! Articles about how to make your code super generic using
category theory are going to alienate people, especially if you are not
*doing* anything with these abilities.