New syntax and semantics for imports
The new syntax has the following variations

    import List
    import List exposing (map, filter)
    import List exposing (..)
    import List as L
    import List as L exposing (..)

In all cases we are able to refer to members of the List module
qualified. In cases 1, 2, and 3, we can say List.foldl or List.isEmpty.
In cases 4 and 5, we give an alias for the module so we can say L.foldl
or L.isEmpty.

When the “exposing” keyword is used, it brings in certain values that
do not need to be qualified. You can explicitly list all exposed
values, but it’s also possible to do a wildcard and just expose
everything.