Properly preserve type aliases after canonicalization and through type inference
When looking for ways to improve performance, look here. Based on the
increased sharing during type inference, this could make things faster.
Based on the use of more data on each value, this could make things
slower.

Could be good to go through and replace Maps with HashMaps and Strings
with Text.