Remove syntax for record insert and remove
These are pretty much never used in practice. They also can lead to
problems if you extend a record with a field that already exists. You
could potentially end up with weird behavior, but since this feature is
rarely used, we have not actually observed this.

Finally, this feature presents an optimization challenge. If the shape
of a record can change as it flows through the program, it may become
more difficult to optimize code to have constant time field lookups and
straightforward field updates.