Avoid cascading errors
When a constraint fails, rather than propagating some bad type through,
just assume the type is “anything”. This means the user will get the
root issue, but not a bunch of stuff that comes from that. Once it is
fixed, they will see any subsequent errors, if there are any.