Generate a minimal number of closures
Optimize out (|>) and (<|)

Introduce concept of Crash so it can be optimized out with the
incomplete pattern match checks (thanks to @JoeyEremondi for this idea!)

Introduce a concept of “Code” in js generation that can be turned into
expressions or statements in a coherent way, allowing us to use
whichever one makes the most sense in a given scenario.

Have if always end with an expression. If it is a multi-if, the last
thing is a crash. The goal would be to optimize it out.

Give specific messages if a condition in an if is not a boolean value.
Explain that there is no truthiness.