Switch Reports to use Docs for pre and post hints
This means we can put all sorts of visual helpers (like colors)
anywhere we want in a hint.

This is the initial pass, getting all the code working in this new
world. Next round will be to hone the usage of colors as I do more
thorough testing of how everything displays.

This also introduces the idea of linking to “hint” pages on certain
errors where the error message could be kind of long.