Remove [1..5] syntax in favor of List.range
The theory is, if you want a range, what do you search for? If you see
[1..5] in the wild, how do you know if it is 1 through 5 inclusive or
exclusive? What do you search for?

The other theory is, if this feature was proposed today, it would
definitely not get added. Used very rarely. The function version is
better. Why special case it? etc.

That said, it does mean our factorial example is a little longer now.

    factorial n =
        List.product [1..n]

    factorial n =
        List.product (List.range 1 n)