Fix #873, detect bad recursive values
This will detect any values defined in terms of themselves without any
lambda in the way to delay evaluation. It cannot catch things like “x =
(\_ -> x) ()” but at that point you are trying to solve the halting
problem, so that is out of scope.

Big thanks to the curious person who inadvertently helped me realize
how to do this!