Create new parser that looks at text directly
This design should generate significantly less garbage.

  - It looks directly at Word16 without creating a Char
  - Variable names are copied out directly, no intermediate linked list
  - Strings are constructed with a Builder in a correctly sized buffer
  - Chars are constructed directly
  - Comments and whitespace don’t allocate anything, just scan words
  - No passes through (>>=) once committed to one of these primitives

I think it will also be possible to avoid a great deal of backtracking
on whitespace. I also suspect this will make it easier to give nice
error messages, but it’ll become clearer when I get to let expressions
and case expressions.