Change how whitespace is checked
Need to produce a column, that way indentation can be checked multiple
times. Needed when you have nested case expressions, so when one fails,
you do not see BeforeIndent on the outer case!

This commit also fixes a number of small errors in the parser:

  - Offset and length are computed correctly for Char
  - Multiline strings work
  - Keywords peek for _ at the end as well as alphaNum characters

All the error messages are still bad, so another pass will be needed to
clean that up!