Improve code that extracts code from source in errors
It now will add the red underline as long as it appears on the last
line of the code snippet. So it’ll be the same for single line errors,
but much nicer for parse errors where the code region spans multiple
lines.

Also switch the public error functions to take a Text version of the
source code.