Use more compact memory representation for Version
This makes the data smaller at runtime (25% of old size + less
indirection) but the real point is to make the binary representation
more compact. In 100% of cases that exist right now, it goes from (3 *
64 = 192) bits to (3 * 8 = 24) bits which is ~12% of the former size.

The assumption here is that no major, minor, or patch will reach 65536
because if you just did patch releases it would take one release per
day for ~180 years.

I would prefer to use Word8 as the runtime representation because 255
seems quite ridiculous as well, but I figured it’d be better to not
have to revisit this for some bizarre case.