Hold less in Interface / add "privatize" to shrink memory
Remove the compiler version and package name from the interface. This
information will be captured by where the interface files live.

The privatize function will drop everything except type information.
This is nice for cases when the module is not actually accessible, but
some of its types may flow into scope anyway.