Refactor representation of "exposing" and documentation
There used to be AST.Variable.Listing, but now everything related to
that lives in AST.Exposing. It also represents the Raw and Canonical
phases, so everything has regions at first, and later it gets nicely
sorted into values, aliases, and unions.

I also cleaned up Docs.AST because I needed to mess with that code
based on the changes in AST.Exposing

Fix #1288, detect when there are duplicates in the module comment.

Also add -Wall and make exports explicit in some cases.