Fix bugs when unifying rigid constraints
The new combineRigidSupers function makes sure that the rigid variable
always wins and that the rigid variable is a subset of the flex
variable.

There was also a bug in unifyRigid where the rigid variable would not
win! The flex one would win. This accounted for some fraction of the
bugs as well.

I also try to avoid allocating new rigid content in unifyRigid to maybe
make things a bit faster.

Fixes #1268
Fixes #1270
Fixes #1281
Fixes #1316
Fixes #1422
Fixes #1581