Disallow equivalent representations of the same type
This is a very roundabout way to fix a big that manifested in elm diff.
When comparing types, some would be generated as (Type “Int”) and
others would be (App (Type “Int”) []) which are the same in meaning,
but different in structure. The core change is to make (Type “Int” [])
the only representation.

Change the binary representation of types. If there is a list of
arguments in (Type name args) we will save ourselves 64 bits whenever
(length args < 256) roughly.

Start using IsString instance of Doc, so (text “(“) becomes “(“ in some
cases.