Make canonicalization hint a bit friendlier
Any level of programmer can be surprised by how `import` works in Elm.
This stuff is always slightly different and it’s always a pain to find
docs on it because it is “so common” to know.

So the new message has the compiler take more responsibility for any
confusion around this.

This change also makes the suggestions yellow.