Keep a pool stack, not just old and young
The Type.State now holds a stack of pools. This way when
Type.Solve.generalize moves things to “old” pools, it can move them as
far away as possible. This means if we are moving from rank 4 to rank
1, we will not need to reconsider the variable again in pools 3 and 2.

This also moves more code from Solver to IO, so there are fewer
liftIO’s sprinkled around.

This reduces the memory usage of the solve phase a decent amount, in my
testing bringing it from ~18% of total allocation down to ~14%