Improve the error messages for bad rigid type variables
Use richer representation of type variables, so FlexVar, FlexSuper,
RigidVar, and RigidSuper are actually different constructors.

This has the downside of adding extra work to pattern matches that
treat all type variables the same. It may be worth considering adding
some indirection for those cases.

It has the benefit of making things very explicit, and the performance
cost appears to be quite small.

The net effect is that error messages with bad rigid variables are much
nicer!