Add intToAscii which converts ints to valid JS field names
It converts an int to ASCII by checking if it is in progressively
larger “block sizes” to make sure we get every possible name.

We start with a block size of 53. This contains [a-zA-Z_]. If that is
too small, we move on to [a-zA-Z_$][a-zA-Z_$0-9] with a block size of
(54 * 64). From there we can just keep getting bigger indefinitely (54
* 64^n) so we never run out of names.

This code also checks for collisions with reserved words in JS,
skipping over them. The logic about block sizes is a bit fancier on
account of this.