Get a decent chunk of expression canonicalization compiling
Some highlights include:

- The “definition sorting” moves into expression canonicalization.
- Cyclic values are not permitted in lets, but cyclic functions are
okay.
- Checks for duplicate record fields are forced by the new AST.
- Local variable shadowing is not allowed. You can shadow foreign stuff
that got in through exposing (..) though.

In 0.18 we would allow cyclic values if the cycle included a lambda. So
if it went through a function or anonymous function, it was permitted
because it MAY actually work. Those sorts of cycles will be permitted
at the top-level though, and the hope is that this is a reasonable
compromise between allowing folks to express everything they need and
avoiding confusing mistakes. It also means that cyclic values can never
be allocated dynamically, which will allow some future GC to not worry
about cyclic data.