Get canonicalization of module going again
This one allows cyclic values in top-level declarations as long as the
cycle includes at least one top-level function as well. This is
slightly more restrictive than the current behavior that allows cycles
as long as it goes through a lambda. That meant that it could use an
anonymous function rather than a named top-level function. I think this
will maintain the ability to express things while catching more
accidental cases.