Implement occurs check with fewer operations and allocations
The old version allocated lists of booleans, then used List.or to check
them all. This meant allocating a bunch of intermediate data structures
and performing a bunch of (||) operations.

The new version is written to thread the accumulated Bool through. This
means there are no intermediate allocations. When a cycle is found, we
can just switch the Bool to True, so there are no (||) operations
either.