Store canonical types with pattern lookup table
This means we will not need to look any types up when generating type
constraints for patterns. Instead of paying O(log2(n)) in the number of
patterns in your dependencies, we pay O(1) to follow a pointer to
things that are already allocated anyway.

I suspect this same strategy can be used for all foreign values as
well. I need to look into the performance of going from a
AST.Type.Canonical to a Type.Type.Variable with the correct rank. My
gut is that it’ll actually be faster than going through Type.Type.Type.