Stop recognizing locally defined binops
Binops can only be used outside of the module they are defined in.

Binops need to work exactly the same as VarForeign, so when that gets
the Type.Canonical added to it, the code in
Canonicalize.Environment.Foreign will need to update for binops as
well. This restriction sets us up for that reality.