Ensure that free type variables are computed only once
This introduces the AST.Canonical.Annotation type that holds on to all
the free type variables in a type. This is then attached to an
VarForeign, VarOperator, or Binop nodes such that we can go from
AST.Canonical.Type directly to a unification variable during type
inference.

Computing and caching the free type variables during canonicalization
means that this work does not need to happen on each use of (|>) or
List.map or whatever else. It may be worth experimenting to see if this
is better overall. Perhaps the memory overhead of having the type
variables for stuff that isn’t used is not worth it.