Revamp type constraint generation
Create new “Expectation” type that holds context information about why
certain things should be equal. That is separate from the “Category”
type that clarifies what kind of node generated the constraint. This
also makes it much easier to “push type information down” into the
branches of ifs and cases.

This also implements constraint generation for non-recursive
definitions. Because we know that they are not self-recursive, we can
generate simpler constraints. Hopefully this speeds things up a bit.