Improve duplicate tracking a bit.
The old version builds Bags as you go, turning it into dictionary at
the end. That means adding things is O(1) and then you pay O(n log(n))
at the end to build the dictionary.

The new one has it as a dictionary all the way through. That means each
insert is O(log(n)) and you end up doing that O(n) times in the worst
case. But you also end up using Dups.union which is faster than
inserting each entry individually.

So the theory is that this should result in a bit less allocation and
be a bit faster.