No more schemes / stop using bags to track pattern bindings
I noticed that every use of CLet had a list like [Scheme] where there
was only one scheme. So rather than doing those two extra allocations
for no reason, I just collapsed them down into the CLet constructor.

This also meant shuffling some code around a bit, from Pattern to
Expression. I think the code is clearer even though it means I have
access to Pattern.State.

I was formerly using bags to track bindings in patterns, but I don’t
think it was saving anything. Normally you are just adding one thing
which is O(1) with lists and bags (though with less allocation in
lists) but sometimes you are adding a list. That case is O(n) for lists
and O(1) for bags, except that the Bag implementation crawled over the
list to turn it into a bag, making it O(n) as well. Then at the end the
bag is converted to a list anyway. So my theory is that going for lists
directly will be faster and allocate less.