Switch canonicalization to CPS / gather usage counts for variables
The main goals here are:

1. Gather usage counts for each argument. This will make it easier to
make inlining decisions in the optimization phase

2. Track whether variable uses are direct or delayed. When deciding if
top-level cycles are valid, this information allows us to accept cycles
of all values as long as the recursion passes through a lambda. Making
this more flexible is based on learnings from writing random Mondrian
generators with Ju. This means it’ll work if the recursion passes
through at least one Random.lazy or Random.andThen.

Switching to CPS makes it easier to do the info tracking IMO, so I just
switched it now.