Find way to quote exported names and make code smaller
In the olden times, we had a series of definitions like this:

var Elm = {};
Elm[‘Thing’] = Elm[‘Thing’] || {};
Elm[‘Thing’][‘Chart’] = Elm[‘Thing’][‘Chart’] || {};
initialize(Elm[‘Thing’][‘Chart’]);

The idea was that Elm was a global object, so you would not want to
clobber things added by other Elm scripts that may be on the page.

I switched things so there is a _Platform_export function that will do
the appropriate kind of export. It looks like this:

_Platform_export({‘Thing’:initialize({‘Chart’:{}})});

Each module name should appear only once, and it is still possible to
expose “mixed-level” modules like Thing.Chart and Thing.Chart.Custom
with this scheme.