Track info for better shadowing error messages
Primarily this starts tracking the location of any given local
definition. That way, if there is any shadowing, I can show the
two-part code snippet where it’s like “I see one definition <HERE> and
another one <THERE>, so which is it?”

It also removes some info from errors when tuples are too large. I
think it makes sense to underline the whole tuple, not just the “extra”
entries. The change is going to be switching to records, so it is not
about removing some entries.