Get cleaner errors on cyclic values
The problem was handling destructuring bindings, like (let (x,y) = … in
…)

Before I generated a unique number for each definition in the let.
Those numbers would be the keys of the graph. After I just use the
variable names as keys, but the destructuring definition turns into a
bunch of nodes. One big node called x$y that has all the real
dependencies, and then a bunch of dummy nodes for the pattern variables
that just point to the big node, like (x => x$y) and (y => x$y)