Try to produce better errors from unification / fixes / optimizations
The root motivation was like this:

1. I want as much control as possible over error messages.
2. The particular “problem” I detect should show up differently in
different circumstances. E.g. I may detect an Int/Float problem, but
the error message is already going to address that in a special way.
3. Detecting “problems” during unification adds a bunch of code to
something that is already pretty complex.
4. I am already crawling these types for the type diff highlighter, so
why not detect the “problems” there?

This change assumes that the answer to (4) is “yeah, that is no
problem!” I suspect this is true.

In stripping the “problem” tracking code out, I improved a couple other
things that I noticed:

- Needed to merge tuples
- Introduced try2 and try3 to unify as much as possible for better
errors
- Made the pattern match for record unification slightly cheaper
- Add unifyArgs to try to do less allocation while being like tryN
- No need to track orientation anymore, slightly reducing allocation.