Change the canonicalization Env for new AST var info
Now that the AST tracks whether something is qualified or not, it makes
sense to change the structure of the canonicalization Env a bit as well.

The new one is focused on finding the good thing withers little extra
mess in the way as possible. If there is an error, I will need to
allocate a bunch of stuff so I can do a more elaborate analysis. It
seems like a fine trade to pay more allocations for quality errors. The
ms spent generating will be worth the minutes saved debugging.

I think it made the code for setting up the environment (in Foreign and
Local) a decent bit simpler. The code cut from Canonicalize.Module is a
good guide to that. I had this two phase thing going for local
declarations and it is not needed it turns out.