Handle qualified/unqualified AST nodes in canonicalization
This commit also introduces AST.Canonical.CtorOpts which caches which
constructor optimizations are possible. The two it allows are for:

type Height = Height Float
type Units = Metric | Imperial

It is fine to unbox Height, so it needs no allocations. It is also fine
to turn Units into a old fashioned enum, where each option is an int,
not an object with an int tag. That one is less obviously helpful, but
it is no extra trouble at this point.