Handle enums and boxed values in optimized AST
The main thing that was tricky here was that the way I destructured
patterns needed to change. Because someone may destructure an argument
like (Parser k) where in PROD it should not turn into a _v4.a access.
There will be no box!

I also treat lists a bit special so I can generate a.b to test if we
are looking at a CONS cell. This works because a.b will be an object in
CONS cells, which turns into true in JS, and it will be undefined in a
NIL, which is false in JS.