Begin revamp of type error messages
There is now a bit more metadata about (1) the expression that
generated the constraint and (2) the context that created the
expectation about that expression.

Based on this information, I made the errors more specific. For
example, when there are problems with (+) or (-) I say something
specifically about numbers, so it is both shorter and more helpful. Or
when the problem is that `func` produces the wrong kind of value, I
really spell that out like “The `func` call is producing:”