Fix bug in decision tree optimizer
The problem was that in the old version, the path to an Alias was
different than a path to other things. So it’d be (Index 2 Alias) vs
(Index 2 Empty) which would avoid collisions with real things that can
be matched on.

This code fixes this not by adding back the “Alias vs Empty”
distinction in paths, but instead by making sure that the alias paths
always come after their content. That means you get to the Ctor before
you get to the Alias.

This change also uses wildcard matches less often. It also fixes some
issues with how tuples were handled.