Create special "error type" that has extra information
The idea is that type inference should produce AST.Canonical.Type if it
succeeds. It can strip out rigid/flex info in that case.

If type inference fails, it can produce an error type that includes
info about infinite loops, errors, rigid variables, etc. I can then
detect problems on that type more easily than doing it during
unification.

This change also introduces some improvements for operator type errors,
specifically on the right side. E.g. if there is an (==) error, the
whole expression should be underlined, not the right, because the error
may be on either side.