Make some progress on type error messages!
The thing I am most excited about is specialized hints that form
learning paths. For example, the hints can now lead you on this path:

‘hello’ + 3
“hello” + 3
“hello” ++ 3
“hello” ++ String.fromInt 3

Or on this one:

3 + userInput
3 + String.toInt userInput
3 + Maybe.withDefault 0 (String.toInt userInput)

I have been wondering “how far can the compiler take a total beginner?”
for a while now, so it is cool to make some progress along those lines!