Switch Error to Exit in builder/
The idea is that compiler/ may produce a bunch of errors. There can be
zero or more of these. That does not mean we have to give up
immediately.

In contrast, builder/ may exit at some point. Something goes wrong and
we just have to stop everything. An exit may contain errors.

This is intended to help with turning exits into JSON in a consistent
way.