Generated type variables are all recorded in Can.FreeVars
If you had a module with unannotated declarations like this:

id v = v

Then it would be exported with an incorrect Can.Annotation because the
generated type variable names (in this case [“a”]) were not recorded
properly.

These changes ensure that the taken names are updated as more names are
generated.