Improve auto-completion a bit more
It turns out it is really complicated to do reasonable things with the
`complete` command.

For example, it is extremely difficult to suggest files yourself
because it will always add a space after directories. If you disable
the adding of spaces, now it does not add spaces after files.

This change makes it possible to get autocompletion set up by putting
the following instruction in your .bash_profile or whatever:

    complete -o default -C elm elm

Elm just knows how to complete itself! So there is no script that has
to go with this.

I also stopped suggesting files, because if you use the -o default
option then it will still suggest files for you. It’s not exactly what
I wanted, but it should work well in any setting where COMP_LINE and
COMP_POINT are set by `complete`