Make sure type information is available for --debug metadata
Before I was trimming interfaces down to only the publicly exposed API.
This meant that .elmi files were smaller, and there was probably some
performance benefit.

BUT I actually need to know about all type shapes in --debug so that
the debugger can tell if your message types have changed. So even if
things are private, the need to get sucked up by that code.