Make Generate.JavaScript.Mode its own module
This change also moves towards getting --debug working again. In
particular, if you are in debug mode, you must have the full transitive
set of module interfaces. By putting that information in the Mode, it
means it only needs to be loaded in that case. In dev and prod mode, no
extra work is needed.