Flow Generate.JavaScript.Mode through other Generate modules
This likely resolves a bug in the old way of generating record fields.
Before we just took what the programmer said, but now we use
Name.fromLocal which will make sure it is not a reserved word.

Probably need a corresponding thing in Debug.toString to remove the _
before these names. Added note about it to my todo list.