type diff implementation with better presentation of aliases
The issue before was that if you had a Model that did not match a { r |
x = t } type, it would unpack both and not tell you anything about
Model anymore.

In the new version, I can say, “I was expecting a `Model` like: { x =
t’ }” so you do not lose that fact.

I can also detect cases where the type alias should not be unpacked. So
if you have an alias of a record, and an alias of an Int, it is
probably easier to understand the issue without unpacking the alias.