promote dependencies properly with `elm install`
Before there was not anything special for the case of promoting
dependencies. For example, it is already in your
“transitive-dependencies” and just needs to move into “dependencies”

The new code detects these cases and gives appropriate information to
the user.

I also made `elm install` always asks for permission to make elm.json
changes. I think it’s just nicer this way.