Try to fix build artifact caching across "git checkout"
Before this commit, we used timestamps to detect if an interface file
(.elmi) needed to be rebuilt. If the .elm file is at t1 and the .elmi
file is at t2, no rebuild is needed if (t1 <= t2)

This did not work if folks did a `git checkout` which switches to a
branch where the .elm file gets older.

So I am trying a new thing. Now .elmi files will have a modification
time exactly equal to the modification time of the corresponding .elm
file. So switching to an older or newer .elm file will always cause a
rebuild.

This can fail is if two people edit the same file at exactly the same
time on different branches. So it is not perfect, but this should
happen less often than the corner case we had before.

The idea is that we will test it out in the pre-release time. Maybe
this trick will not work on certain platforms. We will learn.